Extend the API
