+++ api/app/graphql/types/card_type.rb
@@ -1,0 +1,6 @@
+module Types
+  class CardType < Types::BaseObject
+    field :rank, String, null: false # TODO: move to enum
+    field :color, String, null: true # TODO: move to enum
+  end
+end
