+++ api/app/graphql/types/user_type.rb
@@ -1,0 +1,6 @@
+module Types
+  class UserType < Types::BaseObject
+    field :id, ID, null: false
+    field :name, String, null: false
+  end
+end
